src/backend/src/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
argcomplete==3.5.0
bcrypt==4.2.0
black==24.8.0
certifi==2023.7.22
cffi==1.17.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==43.0.0
datamodel-code-generator==0.25.9
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
genson==1.3.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
inflect==5.6.2
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
jsonpickle==3.2.2
jwcrypto==1.5.6
MarkupSafe==2.1.5
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
phonenumbers==8.13.43
platformdirs==4.2.2
pluggy==1.3.0
psycopg2-binary==2.9.8
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic_core==2.20.1
pytest==7.4.2
pytest-asyncio==0.21.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
StrEnum==0.4.15
toml==0.10.2
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
confluent-kafka
pyyml
pytz
